change type to float

+++ src/vec3/ft_clamp_vec3.c
@@ -6,14 +6,14 @@
 /*   By: wkorande <willehard@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/08 10:55:38 by wkorande          #+#    #+#             */
-/*   Updated: 2020/11/13 23:01:47 by wkorande         ###   ########.fr       */
+/*   Updated: 2020/11/15 11:47:33 by wkorande         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include "vec3.h"
 
-t_vec3	ft_clamp_vec3(t_vec3 v, double min, double max)
+t_vec3	ft_clamp_vec3(t_vec3 v, float min, float max)
 {
 	ft_clamp_d(v.x, min, max);
 	ft_clamp_d(v.y, min, max);
