+++ src/vec3/ft_normalize_vec3.c
@@ -6,7 +6,7 @@
 /*   By: wkorande <willehard@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 14:05:37 by wkorande          #+#    #+#             */
-/*   Updated: 2020/11/13 23:02:27 by wkorande         ###   ########.fr       */
+/*   Updated: 2020/11/15 11:48:12 by wkorande         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 t_vec3	ft_normalize_vec3(t_vec3 v)
 {
 	t_vec3 new;
-	double l;
+	float l;
 
 	l = ft_len_vec3(v);
 	new = ft_div_vec3(v, l);
